# data/test/validation/OpenApiControllerTest_testGetOpenapiDocs.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "GetMyDeck",
    "version" : "v2"
  },
  "servers" : [ {
    "url" : "https://getmydeck.ingenhaag.dev",
    "description" : "production"
  } ],
  "paths" : {
    "/api/v2/summary" : {
      "get" : {
        "tags" : [ "getmydeck-controller" ],
        "operationId" : "getHistoricSummary",
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HistoricSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/regions/{region}/versions/{version}/infos/{reservationTimestamp}" : {
      "get" : {
        "tags" : [ "getmydeck-controller" ],
        "operationId" : "getPersonalInfoResponse",
        "parameters" : [ {
          "name" : "reservationTimestamp",
          "in" : "path",
          "description" : "Time in seconds from 1.1.1970",
          "required" : true,
          "schema" : {
            "minimum" : 1626454800,
            "exclusiveMinimum" : true,
            "maxLength" : 10,
            "minLength" : 10,
            "type" : "string",
            "format" : "[0-9]{10}"
          },
          "example" : 1627022437
        }, {
          "name" : "region",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Region"
          }
        }, {
          "name" : "version",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Version"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Bad request if reservationTimestamp is before official reservation start or in the future"
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InfoResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "HistoricSummary" : {
        "type" : "object",
        "properties" : {
          "summary" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "additionalProperties" : {
                "$ref" : "#/components/schemas/HistoricSummarySet"
              }
            }
          },
          "allDays" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "date"
            }
          }
        }
      },
      "HistoricSummaryEntry" : {
        "type" : "object",
        "properties" : {
          "lastOrderSeconds" : {
            "type" : "integer",
            "format" : "int64"
          },
          "elapsedSeconds" : {
            "type" : "integer",
            "format" : "int64"
          },
          "elapsedDuration" : {
            "type" : "string"
          },
          "lastOrderDate" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "HistoricSummarySet" : {
        "type" : "object",
        "properties" : {
          "entries" : {
            "type" : "object",
            "additionalProperties" : {
              "$ref" : "#/components/schemas/HistoricSummaryEntry"
            }
          }
        }
      },
      "Region" : {
        "type" : "string",
        "enum" : [ "EU", "UK", "US" ]
      },
      "Version" : {
        "type" : "string",
        "enum" : [ "64", "256", "512" ]
      },
      "HistoricDeckbotData" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date",
            "example" : "2022-08-15"
          },
          "elapsedSeconds" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          },
          "elapsedTimePercentage" : {
            "type" : "number",
            "format" : "double",
            "nullable" : true
          },
          "increasedPercentage" : {
            "type" : "number",
            "format" : "double",
            "nullable" : true
          },
          "increasedSeconds" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "InfoResponse" : {
        "type" : "object",
        "properties" : {
          "officialInfo" : {
            "$ref" : "#/components/schemas/OfficialInfo"
          },
          "personalInfo" : {
            "$ref" : "#/components/schemas/PersonalInfo"
          }
        }
      },
      "OfficialInfo" : {
        "type" : "object",
        "properties" : {
          "reservationsStartedAt" : {
            "type" : "string",
            "description" : "UTC time",
            "format" : "date-time",
            "example" : "2021-07-16T17:00:00Z"
          }
        }
      },
      "PersonalInfo" : {
        "type" : "object",
        "properties" : {
          "reservedAt" : {
            "type" : "string",
            "description" : "UTC time",
            "format" : "date-time",
            "example" : "2021-07-23T06:40:37Z"
          },
          "region" : {
            "$ref" : "#/components/schemas/Region"
          },
          "version" : {
            "$ref" : "#/components/schemas/Version"
          },
          "durationReservedAfterStart" : {
            "type" : "string",
            "description" : "String representation of Duration, not yet openapi standard",
            "example" : "PT157H40M37S"
          },
          "durationReservedAfterStartHumanReadable" : {
            "type" : "string",
            "description" : "String representation of Duration, not yet openapi standard",
            "example" : "PT7H40M37S"
          },
          "elapsedTimePercentage" : {
            "type" : "number",
            "format" : "double"
          },
          "prettyText" : {
            "type" : "string"
          },
          "htmlText" : {
            "type" : "string"
          },
          "latestOrderSeconds" : {
            "minimum" : 1626454800,
            "exclusiveMinimum" : true,
            "maxLength" : 10,
            "minLength" : 10,
            "type" : "integer",
            "format" : "int64",
            "example" : 1626819802
          },
          "latestOrder" : {
            "type" : "string",
            "description" : "UTC time",
            "format" : "date-time",
            "example" : "2021-07-23T06:40:37Z"
          },
          "lastDataUpdate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastBatchDate" : {
            "type" : "string",
            "format" : "date",
            "example" : "2022-04-15"
          },
          "historicData" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/HistoricDeckbotData"
            }
          }
        }
      }
    }
  }
}